refactor: Board 방어적 복사 적용
